add users' serializers to serializers.py from folder aerializers
