--- backend/api/serializers/subscription_serializers.py
@@ -1,49 +1,0 @@
-from typing import Dict
-
-from django.contrib.auth import get_user_model
-from rest_framework import serializers
-
-from users.models import Subscription
-from .user_serializers import UserProfileSerializer
-
-User = get_user_model()
-
-
-class SubscriptionSerializer(UserProfileSerializer):
-    """
-    Сериализатор для подписки на автора.
-
-    Атрибуты:
-        email: адрес электронной почты пользователя.
-        username: имя пользователя.
-        first_name: имя пользователя.
-        last_name: фамилия пользователя.
-        is_subscribed: подписка.
-
-    Методы:
-        validate: метод для валидации данных при создании подписки.
-
-    """
-
-    class Meta:
-        model = User
-        fields = ('email', 'id', 'username', 'first_name',
-                  'last_name', 'is_subscribed')
-        read_only_fields = ('email', 'username')
-
-    def validate(self, data: Dict) -> Dict:
-        author = self.instance
-        user = self.context.get('request').user
-        subscription_exists = Subscription.objects.filter(
-            author=author, user=user
-        ).exists()
-
-        if subscription_exists:
-            raise serializers.ValidationError(
-                {'subscription': ['Подписка уже есть']}
-            )
-        if user == author:
-            raise serializers.ValidationError(
-                {'subscription': ['Нельзя подписаться на себя']}
-            )
-        return data
